+++ internal/manager/run.go
@@ -154,7 +154,6 @@ func Run(
 		return fmt.Errorf("failed to create AdminAPIClientsManager: %w", err)
 	}
 	if c.KongAdminSvc.IsPresent() {
-		clients.WithAdminAPIServiceName(c.KongAdminSvc.MustGet())(clientsManager)
 		setupLog.Info("Running AdminAPIClientsManager loop")
 		clientsManager.Run()
 	}